✨ feat: 为删除文章和删除项目添加了鉴权 - 发表文章是允许任何用户发表的

+++ src/hooks/AsyncRequest/Projects/deleteProject.ts
@@ -1,19 +1,26 @@
 import axios from "axios";
 import useRoute from "@/hooks/useRoute.ts";
+import ResponseInfo from "@/type/ResponseInfo.ts";
 
 export default async function (id: string) {
     try {
+        const token = localStorage.getItem('token')
+        if (!token) {
+            const res: ResponseInfo = {
+                status: 402,
+                message: '请先进行登录'
+            }
+            return res
+        }
         const response = await axios({
             method: "DELETE",
             url: `${useRoute.BackEnd}/projects`,
             params: {
                 id,
+                token
             }
         })
-        if (response.data.status == 200) {
-            return '项目删除成功'
-        }
-        return '删除失败，请稍后再试'
+        return response.data
     } catch (err) {
         throw new Error('--服务器状态异常，请稍后再试--');
     }
